text2sql_agent/requirements.txt
streamlit
langchain
langchain-community
langchain-experimental
langchain-google-genai
SQLAlchemy
Faker
python-dotenv
